+++ game/__init__.py
@@ -1,0 +1,2 @@
+COLOR_WHITE = "white"
+COLOR_BLACK = "black"