+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Very summarized. But impressive.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3] - 27-08-2024
+
+### Changed
+
+- Pieces initialization method.
+- Color get method.
+- Piece movement method.
+
 ## [2.2] - 27-08-2024
 
 ### Changed